jmeter and fcm test | fixes

+++ spring-boot-3-jwt-security-main/src/main/java/com/main/repository/ComplaintFileRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface ComplaintFileRepository extends JpaRepository<ComplaintFile, Long> {
-    List<ComplaintFile> findByComplaintId(Integer complaintId);
+    List<ComplaintFile> findByComplaintId(Long complaintId);
 }
